// src/pact-js/pactcore.d.ts
export declare type PluginInteraction = RequestPluginInteraction & ResponsePluginInteraction & RequestResponsePluginInteraction;
export declare type RequestPluginInteraction = {
    withPluginRequestInteractionContents: (contentType: string, contents: string) => boolean;
};
export declare type ResponsePluginInteraction = {
    withPluginResponseInteractionContents: (contentType: string, contents: string) => boolean;
};
export declare type RequestResponsePluginInteraction = {
    withPluginRequestResponseInteractionContents: (contentType: string, contents: string) => boolean;
};
export declare type PluginPact = {
    addPlugin: (plugin: string, version: string) => void;
    cleanupPlugins: () => void;
    cleanupMockServer: (port: number) => boolean;
};
export declare type ConsumerInteraction = PluginInteraction & {
    uponReceiving: (description: string) => boolean;
    given: (state: string) => boolean;
    givenWithParam: (state: string, name: string, value: string) => boolean;
    withRequest: (method: string, path: string) => boolean;
    withQuery: (name: string, index: number, value: string) => boolean;
    withStatus: (status: number) => boolean;
    withRequestHeader: (name: string, index: number, value: string) => boolean;
    withRequestBody: (body: string, contentType: string) => boolean;
    withRequestBinaryBody: (body: Buffer, contentType: string) => boolean;
    withRequestMultipartBody: (contentType: string, filename: string, mimePartName: string) => boolean;
    withResponseHeader: (name: string, index: number, value: string) => boolean;
    withResponseBody: (body: string, contentType: string) => boolean;
    withResponseBinaryBody: (body: Buffer, contentType: string) => boolean;
    withResponseMultipartBody: (contentType: string, filename: string, mimePartName: string) => boolean;
};